requirements.txt
confusable-homoglyphs==3.2.0
Django==2.2.7
django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
django-braces==1.13.0
django-crispy-forms==1.7.2
django-crispy-forms-registration==0.1.3
django-filter==2.2.0
django-modeladmin-reorder==0.3.1
django-registration==3.0.1
django-solo==1.1.3
django-webpack-loader==0.6.0
djangorestframework==3.10.1
gunicorn==19.9.0
psycopg2-binary==2.8.3
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.10.1